src-app/requirements.txt
flask>=2.0.0
pyyaml>=5.0.0
boto3>=1.26.0
botocore>=1.29.0
